requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.9
Django==4.0
django-ckeditor==6.2.0
django-js-asset==1.2.2
django-taggit==2.0.0
gunicorn==20.1.0
idna==3.3
Pillow==8.4.0
requests==2.26.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.7
whitenoise==5.3.0
